:recycle: Refactor transaction modifier
